+++ spec/rubocop/cop/style/method_call_parentheses_spec.rb
@@ -6,7 +6,7 @@ module Rubocop
   module Cop
     module Style
       describe MethodCallParentheses do
-        let(:cop) { MethodCallParentheses.new }
+        let(:cop) { described_class.new }
 
         it 'registers an offence for parens in method call without args' do
           inspect_source(cop, ['top.test()'])
@@ -18,6 +18,11 @@ module Rubocop
 
         it 'it accepts parens in method call with args' do
           inspect_source(cop, ['top.test(a)'])
+        end
+
+        it 'auto-corrects by removing unneeded braces' do
+          new_source = autocorrect_source(cop, 'test()')
+          expect(new_source).to eq('test')
         end
       end
     end